used only router
