+++ wolox-books/src/app/screens/auth/auth.module.ts
@@ -2,14 +2,14 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AuthComponent } from './auth.component';
 import { BookListModule } from './screens/book-list/book-list.module';
-import { AuthRoutingModule } from './auth-routing.module';
+import { AppRoutingModule } from 'src/app/app-routing.module';
 
 @NgModule({
   declarations: [AuthComponent],
   imports: [
     CommonModule,
     BookListModule,
-    AuthRoutingModule,
+    AppRoutingModule
   ],
   exports: [AuthComponent]
 })
